apps/api/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.1

# Background Tasks
celery==5.3.6
arq==0.25.0

# Twitter API
httpx==0.26.0
tweepy==4.14.0
authlib==1.3.0

# AI
anthropic==0.18.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.1.0

# Utils
python-slugify==8.0.1
python-dotenv==1.0.0

# WebSocket
websockets==12.0

# Image Generation (OG images)
Pillow==10.2.0

